+++ src/App.vue
@@ -12,6 +12,7 @@
     <ProjectConfigModal />
     <SshConfigModal />
     <GitModal />
+    <IntroModal />
   </div>
 </template>
 
@@ -25,6 +26,7 @@ import SpecialPushModal from './components/modals/SpecialPushModal.vue';
 import ProjectConfigModal from './components/modals/ProjectConfigModal.vue';
 import SshConfigModal from './components/modals/SshConfigModal.vue';
 import GitModal from './components/modals/GitModal.vue';
+import IntroModal from './components/modals/IntroModal.vue';
 
 import { useProjects } from './composables/useProjects';
 import { useSsh } from './composables/useSsh';